♻️ Assets/Tabs 图标合并进 Sidebar，移除独立 ActivityBar
